refactor: apply header changes to source files
