+++ src/net/session.c
@@ -24,11 +24,15 @@
 #include "zenoh-pico/net/memory.h"
 #include "zenoh-pico/protocol/core.h"
 #include "zenoh-pico/session/utils.h"
-#include "zenoh-pico/transport/link/task/join.h"
-#include "zenoh-pico/transport/link/task/lease.h"
-#include "zenoh-pico/transport/link/task/read.h"
+#include "zenoh-pico/transport/common/join.h"
+#include "zenoh-pico/transport/common/lease.h"
+#include "zenoh-pico/transport/common/read.h"
 #include "zenoh-pico/transport/multicast.h"
+#include "zenoh-pico/transport/multicast/lease.h"
+#include "zenoh-pico/transport/multicast/read.h"
 #include "zenoh-pico/transport/unicast.h"
+#include "zenoh-pico/transport/unicast/lease.h"
+#include "zenoh-pico/transport/unicast/read.h"
 #include "zenoh-pico/utils/logging.h"
 #include "zenoh-pico/utils/uuid.h"
