+++ src/transport/multicast/join.c
@@ -12,10 +12,10 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
-#include "zenoh-pico/transport/link/task/join.h"
+#include "zenoh-pico/transport/multicast/join.h"
 
 #include "zenoh-pico/session/utils.h"
-#include "zenoh-pico/transport/link/tx.h"
+#include "zenoh-pico/transport/multicast/tx.h"
 
 #if Z_FEATURE_MULTICAST_TRANSPORT == 1
