+++ src/transport/transport.c
@@ -21,10 +21,12 @@
 #include "zenoh-pico/config.h"
 #include "zenoh-pico/link/link.h"
 #include "zenoh-pico/protocol/core.h"
-#include "zenoh-pico/transport/link/rx.h"
-#include "zenoh-pico/transport/link/tx.h"
-#include "zenoh-pico/transport/multicast.h"
-#include "zenoh-pico/transport/unicast.h"
+#include "zenoh-pico/transport/multicast/rx.h"
+#include "zenoh-pico/transport/multicast/transport.h"
+#include "zenoh-pico/transport/multicast/tx.h"
+#include "zenoh-pico/transport/unicast/rx.h"
+#include "zenoh-pico/transport/unicast/transport.h"
+#include "zenoh-pico/transport/unicast/tx.h"
 #include "zenoh-pico/transport/utils.h"
 #include "zenoh-pico/utils/logging.h"
